# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Sf Academy
  version: 1.0.0
  description: Generated docs from my candidacy test
paths:
  /exchange:
    get:
      description: Converts to other currency
      operationId: exchange
      x-eov-operation-handler: routes/exchange
      parameters:
        - name: value
          in: query
          required: true
          schema:
            type: number
            minimum: 1
            maximum: 1000000
            nullable: false
        - name: from
          in: query
          schema:
            allOf:
              - $ref: "./schemas/currency.yaml"
        - name: to
          in: query
          schema:
            allOf:
              - $ref: "./schemas/currency.yaml"
      responses:
        "200":
          description: Succesful exchange
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ./schemas/amount.yaml
  /login:
    post:
      description: Release a JWT on succesful login
      operationId: login
      x-eov-operation-handler: routes/login
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "./schemas/userCredentials.yaml"
      responses:
        "201":
          description: Succesful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "./schemas/errorResponse.yaml"
  /signup:
    post:
      description: Inserts new user on succesful signup
      operationId: signup
      x-eov-operation-handler: routes/signup
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "./schemas/userData.yaml"
      responses:
        "201":
          description: Succesful signup
          content:
            application/json:
              schema:
                $ref: "./schemas/okResponse.yaml"
        "409":
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: "./schemas/errorResponse.yaml"
  /deposit:
    post:
      description: Create a deposit-like transaction
      operationId: deposit
      x-eov-operation-handler: routes/deposit
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "./schemas/amount.yaml"
      responses:
        "201":
          description: Succesful deposit
          content:
            application/json:
              schema:
                $ref: "./schemas/okResponse.yaml"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "./schemas/errorResponse.yaml"
      security:
        - bearerAuth: []
  /withdraw:
    post:
      description: Create a withdraw-like transaction
      operationId: withdraw
      x-eov-operation-handler: routes/withdraw
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "./schemas/amount.yaml"
      responses:
        "201":
          description: Succesful withdraw
          content:
            application/json:
              schema:
                $ref: "./schemas/okResponse.yaml"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "./schemas/errorResponse.yaml"
      security:
        - bearerAuth: []   
  /listTransactions:
    get:
      description: Returns the list of transactions of authenticated user
      operationId: listTransactions
      x-eov-operation-handler: routes/listTransactions
      responses:
        "200":
          description: Succesful list
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "./schemas/transaction.yaml"
                  required:
                    - eurDelta
                    - usdDelta
                    - timestamp
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "./schemas/errorResponse.yaml"
      security:
        - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT